requirements.txt
tweepy==4.14.0
python-telegram-bot==20.6
linkedin-api==2.0.0
openai==1.3.0
anthropic==0.8.0
transformers==4.35.0
torch>=2.1.0
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
newsapi-python==0.2.7
schedule==1.2.0
pandas==2.1.3
numpy==1.25.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.0
asyncio==3.4.3
loguru==0.7.2
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
